requirements.txt
# Google Cloud
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
db-dtypes>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Utils
python-dateutil>=2.8.0
tqdm>=4.65.0
